+++ RobBiped/RobBiped/Kinematics_Dynamics/CoM_location.cpp
